+++ src/main/java/ch/njol/skript/expressions/ExprRound.java
@@ -72,10 +72,10 @@ public final class ExprRound extends PropertyExpression<Number, Long> {
 			@Override
 			public Long convert(final Number n) {
 				if (n instanceof Integer)
-					return Long.valueOf(n.longValue());
+					return n.longValue();
 				else if (n instanceof Long)
 					return (Long) n;
-				return Long.valueOf(action == -1 ? Math2.floor(n.doubleValue()) : action == 0 ? Math2.round(n.doubleValue()) : Math2.ceil(n.doubleValue()));
+				return action == -1 ? Math2.floor(n.doubleValue()) : action == 0 ? Math2.round(n.doubleValue()) : Math2.ceil(n.doubleValue());
 			}
 		});
 	}
